add test
remove the un-used packages
