--- production/dea_ml/docker/requirements-jupyter.txt
@@ -1,56 +1,0 @@
-# jupyter things
-jupyter==1.0.0
-jupyterlab==2.2.9
-jupyterlab_iframe==0.2.2
-jupyterhub==1.3.0
-ipywidgets==7.5.1
-ipyleaflet==0.13.0
-jupyter-server-proxy==1.5.0
-nbdime==2.0.0
-jupyterlab-code-formatter==1.3.6
-sidecar==0.4.0
-ipyevents==0.7.1
-ipycanvas==0.4.6
-ipyfilechooser==0.3.4
-jupyterlab-git==0.20.0
-nbgitpuller
-jupyter-nbextensions-configurator
-jupyter-contrib-nbextensions
-jupyter-contrib-core
-jupyter-ui-poll
-nbresuse==0.3.3
-black
-autopep8
-yapf
-isort
-mypy
-line_profiler
-nbconvert==5.6.1
-nbformat==5.0.8
-nbval==0.9.6
-jupytext==1.5.2
-# fix for auto-complete breaking with jedi 0.18
-jedi==0.17.2
-parso==0.7.1
-
-# jupyter + matplotlib
-# matplotlib itself is in requirements.txt as it's pulled in by other libs there
-ipympl==0.5.6
-graphviz
-folium
-pydotplus
-plotly
-
-# jupyter + bokeh
-#
-jupyter-bokeh==2.0.2
-dask-labextension==2.0.2
-geoviews==1.8.1
-cartopy==0.18.0
-datashader==0.11.0
-panel==0.9.7
-holoviews==1.13.2
-
-# jupyterhub + dask
-dask-gateway==0.9.0
-dask-gateway-server==0.9.0
